Initial channel implemtation
The HalObject class is added here which abstracts the idea of a module
or an agent. HalModule and HalAgent only differ in send behavoir, more
specifically send behavoir when out channels are not specified.

Module/agent instance configs can now contain an 'out' field which is a
list of names of objects to send the output of the module to (via send).
The absence of this invokes the default behavoir (send to all modules
for agents, the agent specified in the message for modules).