loader: implement new format for loading module classes
Previously, the loader scanned loaded module files for classes
that inherited the halibot.HalModule class. This was not only
inefficent, but also buggy with further subclasses (e.g.
halibot.CommandModule) and rather unpythonic.

The new specifier in a module's config.json is as follows:
	"main": "file.py:ModuleName"

The loader will attempt to split on ':', and if successful,
load the first as the module file, and the latter of a HalModule.
There is no longer any silly type checking, so not inheriting from
HalModule is totally possibly, but probably not a good idea.

The old loading style still remains for if the 'main' key can't be
split, but should probably be removed in the future.