Some minor overhauls (#40)
* loader: implement new format for loading module classes

Previously, the loader scanned loaded module files for classes
that inherited the halibot.HalModule class. This was not only
inefficent, but also buggy with further subclasses (e.g.
halibot.CommandModule) and rather unpythonic.

The new specifier in a module's config.json is as follows:
	"main": "file.py:ModuleName"

The loader will attempt to split on ':', and if successful,
load the first as the module file, and the latter of a HalModule.
There is no longer any silly type checking, so not inheriting from
HalModule is totally possibly, but probably not a good idea.

The old loading style still remains for if the 'main' key can't be
split, but should probably be removed in the future.

* main: remove useless 'imp' import

* main: add command line arguments, add some instance init

Slight overhaul of the main halibot entry point. Includes the following
arguments:
	init - Create a template config.json, modules, and agents dir
	run  - Run a halibot instance using the local dir's config
	help - Show the help.

* Update to use new module/agent config format

Update configs of:
 * irc
 * agentcontainer
 * modulecontainer

* Change levels of some loader logs

Things that cause a module to not be loaded should be an error, as
Halibot will not behave as expected (i.e. use the specified module) in
those cases. Also, non-breaking behavoir that is deprecated and can
therefore be removed in the future should be a warning.