Implement resource identifiers
Contexts have been deprecated. Instead of a context, each message has an
'origin' which is a resource identifier (RI) that indicates where the
message is from. RIs are of the form 'name/resource' where 'name'
identifies the module or agent and resource is object-specific, e.g. in
the IRC agent the resource is the channel or user communicated to/from.

RIs should be used in place of context objects in configs and elsewhere,
in theory drastically simplifying configs.

The use of contexts is still supported for backwards compatibility.
Support for contexts will be removed in 1.0.