halibot: integrate halauth into halibot core
Add the HalAuth object into Halibot core.

An optional config key has been added: "auth-path". This points to a particular
file which holds the permissions list. If not supplied, the default path of
"$CURRENTDIR/permissions.json" is used.

HalAuth can be disabled either by setting the "use-auth" key to False.Disabling
auth puts the HalAuth object in a permissive mode, that always returns True.
This is currently for legacy compatibility, but should probably be configurable.