Initial package implementation
The only breaking change that results from this is that modules and
agents are now specified with a ':ClassName' following the package name
(e.g. 'irc:IrcAgent').

There is also a new external function on the Halibot object,
'get_package', which returns the python module object associated with
the given package name. Non-relative package paths may not work as
expected.

Backwards compatibility with old-style modules and agents is maintained,
to be removed in 1.0.