message: convert author to a dynamic attribute
A message's author was a flat string, that the agent sets for modules to
read and react to. This changes the behavior of the "author" attribute to
allow for extra data to be stored in there seamlessly from the original
purpose.

This can be used to embed unique identifying information decoupled from the
nickname of the user, as well as a long name vs short (sensible) name.

Example:
>>> msg.author
"richteer"
>>> msg.author.id
"richteer@lyra.lastprime.net"

Or when there is no unique identifying information:
>>> msg.author = "foobar"
>>> msg.author
"foobar"
>>> msg.author.id
""