Initial Implementation of Basic Auth (#42)
* halauth: add authentication object

HalAuth is a new object intended to sit in Halibot core, and handles permission
objects.

Authentication objects currently take the form of a triple:
(RI, identity, Permission String)
...which binds a particular permission to a identity and RI.

An example would be the following:
(irc/##halibot, richteer, CanReloadModules)
...which essentially says that the user "richteer" in the IRC channel "##halibot"
is allowed to operate anything that requests the "CanReloadModules" permission.

Any of the three can be wildcarded, including all of them at once. Currently,
the wildcard indicator is an asterisk ("*"). This indicates that ANY of that
particular field matches, so allowing all permissions to a particular user in
a particular channel (RI) would look like this:
(*, richteer, *)

The current API on the HalAuth module includes three methods:
 - grantPermission(ri, identity, perm) -> add a perm triple to the list
 - revokePermission(ri, identity, perm) -> remove an EXACT MATCHING triple
 - hasPermission(ri, identity, perm) -> check if a triple matches one in the list

* halibot: integrate halauth into halibot core

Add the HalAuth object into Halibot core.

An optional config key has been added: "auth-path". This points to a particular
file which holds the permissions list. If not supplied, the default path of
"$CURRENTDIR/permissions.json" is used.

HalAuth can be disabled either by setting the "use-auth" key to False.Disabling
auth puts the HalAuth object in a permissive mode, that always returns True.
This is currently for legacy compatibility, but should probably be configurable.

* halmodule: add hasPermission helper func for module developers

Add a simple wrapper to make the call to HalAuth's hasPermission().

This implementation takes in a message object for simplicity.

* tests: add initial halauth testing, only hasPermission methods for now

* halauth: add @hasPermission() module method decorator

Implements a function decorator for easy permission checking at
the method-level.

The decorater takes in a permission string as an argument, and expects
to be attached to a method with a message object as the first argument.

Example module usage: (and main use-case):

```
class ExampleModule

	@hasPermission("CanCommand")
	def SomeCommand(self, msg):
		# protected code here...

	def receive(self, msg):
		if msg.body.startswith("!command"):
			self.SomeCommand(msg)
```

* message: add identity field to messages

Adds a field for the author's uniquely identifying string, for purposes such as
permissions, and per-user data storage.