Merge add-agent and add-module commands into add
Now there is just a single add command that adds modules and agents.
Should certainly be easier on the user. It detects whether to add it as
an module or agent be checking if the class to add is a subclass of the
module or agent class. There is also a command line argument to override
this decision or to make it in the case where neither are inherited.
Uses for those options ('-a' and '-m') should probably be rare unless
you are doing something funky though.