halobject: add module interactivity helper via invoke()
For more information on the usage of this, see
doc/ModuleInteractivity.md

This commit introduces the simplest top-level implementation of module
interactivity, by providing a method to hide away the needed calls
to Halibot core.

Further wrappers can be built around this, as it intentionally
propogates up exceptions if A) the module in question does not exist or
B) the method does not exist on the target (or if the arguments don't
match...).