halibot: merge the agents and modules dict into one (#61)
* halibot: merge the agents and modules dict into one

For a long time, it has been debated if the agents and modules
dictionaries really needed to be separate. Both contain HalObjects with
similar interfaces, so they are not so different. However, with the
introduction of RIs, the leading argument against merging them has been
eliminated: agent and module instances can no longer share the same
name.

Since an agent and module can't have the same name, there is effectively
no longer a point to separating them. They have been replaced by a small
dictionary class that provides read-only attributes that handle the
filtering for you: .agents and .modules for geting just agents and
modules respectively.

This simplifies the lookup and addition of new objects, as there is no
longer two separate dictionaries that need to be examined.

* halibot: remove get_object method

Since agents and modules have been merged into one dictionary, this
patch removes the get_object() method, and replaces it with a direct
call to the objects' get() method.

* halibot: merge loading functions

Continuing on the trend of merging duplicate efforts, this commit
replaces add_*_instance with add_instance, since it is completely
independent of what type of object is being added.

The internal loading function for loading modules/agents from the config
has also been merged into _instantiate_objects, which now takes which
objects to load.