halmodule: fix false origin on replies
Origin should be set to wherever a message comes from -- this includes
modules. Therefore, we should not be setting the origin to where the message
we are replying to is from.

e.g. replying to a message from "irc/##foobar" should not reply with the
origin "irc/##foobar" as well. This could cause problems with replying
to replies.