halconfigurer: use proxy to call input() in configure() step
Instead of using input() directly, use a separate overrideable function so
that the input could be pulled from another source instead.

This adds a nice entry point for unit testing to force particular inputs
without having to overhaul the entire source. Also, in theory this permits
a function to be injected that uses .sync_send_to(), so a config module
can interactively reconfigure another module interactively through an agent.