halauth: rework the hasPermission decorator for better flexibility
The `@hasPermission` decorator was intended to be simple to use for module
developers. As it turns out however, it is completely incompatible with modules
that subclass CommandModule. Considering this is also a class intended to make
life easier, something had to be fixed.

Thus, the decorator now supports two optional parameters when applying to a
function: "argnum" and "key". Since we need to extract the source origin and
identity at runtime, but different modules may have different calling
conventions, this was the easiest and most flexible method to instruct the
decorator where to look for its information.

argnum can be supplied to specify which positional argument contains a Message
object, indexing by 0. Remember, class methods do not count self in *args!.
The default value is None, so unless supplied, this parameter is ignored.

key can be supplied to specify a key for retrieveing the Message object from
**kwargs. The default value is "msg", which is the default convention for
classes that inherit from CommandModule.

If argnum is specified (i.e. not None), it takes priority over key.

Example:

class MyModule(HalModule):
	# Uses first non-self positional argument
	@hasPermission("FOO", argnum=0)
	def myfunc(self, msg):
		pass

	# key= is not necessary here, since "msg" is the default
	#  make sure msg= is supplied!
	@hasPermission("BAR")
	def mybar(self, msg=None):
		pass