halauth: simplify dependencies for the hasPermission decorator
The `@hasPermission` decorator was intended to be simple to use for module
developers. As it turns out however, it is completely incompatible with modules
that subclass CommandModule. Considering this is also a class intended to make
life easier, something had to be fixed.

Thus, a compromise was reached: have the decorator search for the `Message`
object *somewhere* in the target function's parameters. Rather than require it
to be the first positional argument, it will instead use the identity/ri
information from the *first* `Message` object in the parameter list.