Implement local/system configs
System config lives in the source directory. Added a special variable to
halibot configs. If this was not done, then there would no way to keep
the system config pointing to the source package path without doing some
subsitution or something in install.sh, which is not ideal as it would
not work out-of-the-box, only after installation.