halobject: split raw sending logic into its own function
Functions like .send_to() really should be a helper function that sets the
target of a message for you (and handles multiple targets), but it shouldn't
be the bottom sending function itself.

This commit moves the actual send logic (including copy) out of the .send_to
function, and into its own function called .raw_send(). A `Message` object
parameter MUST be populated with a valid `.target` and `.origin` parameter.