halauth: add ability for hasPermission to still reject a call if permissions are disabled
Some functionalities really shouldn't be exposed without permission. This
commit adds the "permissive" option to hasPermission (and the decorator)
to declare what the default behavior should be if halibot auth is disabled.

This will prevent admin commands like "!shutdown" being run if permissions
are not enabled -- you must explicitly enable permissions and grant the
proper user permissions to actually use that command.