commandmodule: add namespacing feature, for simple subcommands
A common pattern is to have one !modname command and have a bunch of
subcommands branch from that, e.g. !modname foo . This leads to the
command table of the CommandModule being relatively small, while
putting a lot of repeated effort in subcommand parsing.

This commit introduces a new optional property "namespace", that acts
like a command prefix -- if set, all commands in the table are treated
as subcommmands to a root !command.

For example, the Admin core module sets namespace "admin", so its commands
"reload", and "shutdown" must be called like so:
!admin reload
!admin shutdown.

Since the namespace is evaluated at runtime, the namespace can be set by
the module, module configuration, or even at runtime by another module.