commandmodule: add reason for command rejection, factor out default handling
With the addition of namespaces, the number of cases to exit early from
command receiving is higher, and also increases the number of places to
call the default handler.

This commit factors out the command handling into its own function, and calls
the default handler based on the return of the command handler. The command
handler now returns an enum (or None) stating the reason why the command
was rejected. This reason is now passed as an optional argument to the
default handler for the module to handle without needing to re-parse the
command to see why.